requirements.txt
polars==1.29.0
python-dateutil==2.9.0.post0
streamlit==1.45.0
